feat:add asset log details
